requirements.txt
Django==5.1.2
requests==2.26.0
python-decouple